- user: заменил загрузку целого User в delete() на явную   проверку existsById() - item: убрал приватный метод validateForCreate()

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -55,7 +55,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long userId) {
-        getUserOrThrow(userId);
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        }
         userRepository.deleteById(userId);
     }
